chore: promotion type 1 implemetation

+++ ShoppingCart/Models/ShoppingItem.cs
@@ -8,6 +8,7 @@ namespace ShoppingCart.Models
     {
         public SKU Item { get; set; }
         public int Quantity { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
+
     }
 }
